Remove :applications instructions from readme
Elixir 1.4 (released in 2016) add application inference, so now in
almost all situations it is not necessary to add a dependency to
`:applications`. And if you set just a single application in
`:applications` then you will break your release.

Since ueberauth_google only support Elixir 1.8 and up, we don't need to
mention :applications anymore